requirements.txt
# Core dependencies
python-dotenv==0.19.2
python-docx==0.8.11
pydantic==1.10.24

# Document processing
openpyxl==3.0.10
python-pptx==0.6.21
PyPDF2==3.0.1

# Vector store and embeddings (optional - requires Python 3.8+)
# chromadb==0.3.29
# sentence-transformers==2.2.2

# Web scraping
requests==2.27.1
beautifulsoup4==4.10.0

# LLM APIs
openai==0.28.1

# Search API
tavily-python==0.1.9

# CLI and UI
streamlit==1.28.0
